# 99_MyFiles/Project_Trial/Challenge_5_v2/contracts/api-contract.yaml
openapi: 3.1.0
info:
  title: SharkNinja Consumer Sentiment Dashboard
  description: Consumer-electronics-aware NLP pipeline for social listening. Handles
    sarcasm, comparative sentiment, ABSA, and CE domain terminology.
  version: 1.0.0
paths:
...
  /api/products:
    get:
      tags:
      - products
      summary: List Products
      description: List all products with summary stats (brand, category, overall
        score, mention count).
      operationId: list_products_api_products_get
      parameters:
      - name: brand
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Brand'
          - type: 'null'
          title: Brand
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
          title: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductSummary'
                title: Response List Products Api Products Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/products/{product_model}/aspects:
    get:
      tags:
      - products
      summary: Get Product Aspects
      description: Per-product aspect breakdown with trend and severity.
      operationId: get_product_aspects_api_products__product_model__aspects_get
      parameters:
      - name: product_model
        in: path
        required: true
        schema:
          type: string
          title: Product Model
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductAspectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/products/{product_model}/timeseries:
    get:
      tags:
      - products
      summary: Get Product Timeseries
      description: Per-product daily sentiment timeseries.
      operationId: get_product_timeseries_api_products__product_model__timeseries_get
      parameters:
      - name: product_model
        in: path
        required: true
        schema:
          type: string
          title: Product Model
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeseriesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/products/{product_model}/forecast:
    get:
      tags:
      - products
      summary: Get Product Forecast
      description: '4-week forward sentiment forecast using exponentially-weighted
        linear projection.


        Use ?explain=true to return the formula documentation instead of forecast
        data.


        Method: heuristic projection, NOT an LLM simulation. See forecast.py for full
        formula.

        low_confidence=true when mention_count < 50 OR window_days < 14.'
      operationId: get_product_forecast_api_products__product_model__forecast_get
      parameters:
      - name: product_model
        in: path
        required: true
        schema:
          type: string
          title: Product Model
      - name: explain
        in: query
        required: false
        schema:
          type: boolean
          description: Return formula documentation instead of forecast data
          default: false
          title: Explain
        description: Return formula documentation instead of forecast data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/topics:
    get:
      tags:
      - topics
      summary: Get Topics
      description: 'Emerging topics clustered by aspect, sorted by momentum.

        Each topic includes exemplar quotes for evidence traceability.'
      operationId: get_topics_api_topics_get
      parameters:
      - name: brand
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Brand'
          - type: 'null'
          title: Brand
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
          title: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicExplorerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/topics/comparative:
    get:
      tags:
      - topics
      summary: Get Comparative Topics
      description: 'Share-of-aspect comparison between two brands.

        Shows which aspects each brand dominates in consumer conversation.'
      operationId: get_comparative_topics_api_topics_comparative_get
      parameters:
      - name: brand_a
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Brand'
          description: First brand for comparison
        description: First brand for comparison
      - name: brand_b
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Brand'
          description: Second brand for comparison
        description: Second brand for comparison
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
          title: Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparativeTopicResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AspectSentiment:
      properties:
        name:
          type: string
          title: Name
          description: Canonical aspect name, e.g. 'suction', 'battery', 'brushroll'
        polarity:
          $ref: '#/components/schemas/Polarity'
        score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Score
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
        snippet:
          type: string
          title: Snippet
          description: Text fragment that triggered this aspect extraction
      type: object
      required:
      - name
      - polarity
      - score
      - confidence
      - snippet
      title: AspectSentiment
    AspectTrend:
      properties:
        aspect:
          type: string
          title: Aspect
          description: Canonical aspect name
        mention_count:
          type: integer
          title: Mention Count
        avg_score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Avg Score
        trend_delta:
          type: number
          title: Trend Delta
          description: 30-day score change
        severity:
          type: number
          title: Severity
          description: volume x magnitude x recency composite
        sparkline:
          items:
            type: number
          type: array
          title: Sparkline
          description: Weekly scores, most recent last
      type: object
      required:
      - aspect
      - mention_count
      - avg_score
      - trend_delta
      - severity
      - sparkline
      title: AspectTrend
    Brand:
      type: string
      enum:
      - shark
      - ninja
      - dyson
      - irobot
      - roborock
      - kitchenaid
      - breville
      - cuisinart
      - keurig
      - delonghi
      - other
      title: Brand
...
    Category:
      type: string
      enum:
      - robot_vacuum
      - cordless_stick
      - upright
      - air_fryer
      - pressure_cooker
      - blender
      - ice_cream_maker
      - coffee
      - air_purifier
      - hair_tool
      - other
      title: Category
    ComparativePair:
      properties:
        brand:
          $ref: '#/components/schemas/Brand'
        aspect:
          type: string
          title: Aspect
          description: Canonical aspect name
        polarity:
          $ref: '#/components/schemas/Polarity'
        score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Score
      type: object
      required:
      - brand
      - aspect
      - polarity
      - score
      title: ComparativePair
    ComparativeTopicResponse:
      properties:
        brand_a:
          $ref: '#/components/schemas/Brand'
        brand_b:
          $ref: '#/components/schemas/Brand'
        share_of_aspect:
          items:
            $ref: '#/components/schemas/ShareOfAspect'
          type: array
          title: Share Of Aspect
      type: object
      required:
      - brand_a
      - brand_b
      - share_of_aspect
      title: ComparativeTopicResponse
    DerivedSentiment:
      properties:
        overall_sentiment:
          $ref: '#/components/schemas/SentimentLabel'
        compound_score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Compound Score
          description: Single scalar for charting
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          description: Pipeline confidence in this result
        sarcasm_flag:
          type: boolean
          title: Sarcasm Flag
          description: True if sarcasm heuristics fired
        aspects:
          items:
            $ref: '#/components/schemas/AspectSentiment'
          type: array
          title: Aspects
        comparative_pairs:
          anyOf:
          - items:
              $ref: '#/components/schemas/ComparativePair'
            type: array
          - type: 'null'
          title: Comparative Pairs
          description: Only populated when a comparative claim is detected
      type: object
      required:
      - overall_sentiment
      - compound_score
      - confidence
      - sarcasm_flag
      title: DerivedSentiment
    ForecastPoint:
      properties:
        date:
          type: string
          title: Date
          description: ISO date YYYY-MM-DD
        projected_score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Projected Score
        confidence_lower:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Confidence Lower
        confidence_upper:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Confidence Upper
      type: object
      required:
      - date
      - projected_score
      - confidence_lower
      - confidence_upper
      title: ForecastPoint
    ForecastResponse:
      properties:
        product_model:
          type: string
          title: Product Model
        historical:
          items:
            $ref: '#/components/schemas/TimeseriesPoint'
          type: array
          title: Historical
        forecast:
          items:
            $ref: '#/components/schemas/ForecastPoint'
          type: array
          title: Forecast
        method_label:
          type: string
          title: Method Label
          description: Human-readable label for the forecasting method used
        input_mention_count:
          type: integer
          title: Input Mention Count
        input_window_days:
          type: integer
          title: Input Window Days
        low_confidence:
          type: boolean
          title: Low Confidence
          description: True when mention_count < 50 OR window_days < 14
        caveats:
          items:
            type: string
          type: array
          title: Caveats
          description: Reasons for low confidence or forecast limitations
      type: object
      required:
      - product_model
      - historical
      - forecast
      - method_label
      - input_mention_count
      - input_window_days
      - low_confidence
      title: ForecastResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Mention:
      properties:
        mention_id:
          type: string
          title: Mention Id
          description: Stable UUID identifier
        source_platform:
          $ref: '#/components/schemas/SourcePlatform'
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
          description: URL to original mention; null for closed sources
        author_handle:
          anyOf:
          - type: string
          - type: 'null'
          title: Author Handle
          description: Anonymized/hashed author handle
        posted_at:
          type: string
          format: date-time
          title: Posted At
          description: When the mention was published (UTC)
        ingested_at:
          type: string
          format: date-time
          title: Ingested At
          description: When our pipeline processed it (UTC)
        brand:
          $ref: '#/components/schemas/Brand'
        category:
          $ref: '#/components/schemas/Category'
        product_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Model
          description: Specific SKU name if available
        text:
          type: string
          title: Text
          description: Raw mention text, pre-NLP
        rating:
          anyOf:
          - type: number
          - type: 'null'
          title: Rating
          description: Source-provided star rating if available
        language:
          type: string
          title: Language
          description: ISO-639-1 language code
          default: en
        derived:
          $ref: '#/components/schemas/DerivedSentiment'
        record_type:
          type: string
          enum:
          - review
          - recall
          title: Record Type
          description: '''review'' for normal mentions; ''recall'' for CPSC recall
            records'
          default: review
      type: object
      required:
      - mention_id
      - source_platform
      - posted_at
      - ingested_at
      - brand
      - category
      - text
      - derived
      title: Mention
...
    Polarity:
      type: string
      enum:
      - positive
      - negative
      - neutral
      - mixed
      title: Polarity
    ProductAspectResponse:
      properties:
        product_model:
          type: string
          title: Product Model
        brand:
          $ref: '#/components/schemas/Brand'
        category:
          $ref: '#/components/schemas/Category'
        aspects:
          items:
            $ref: '#/components/schemas/AspectTrend'
          type: array
          title: Aspects
        exemplar_mentions:
          items:
            $ref: '#/components/schemas/Mention'
          type: array
          title: Exemplar Mentions
      type: object
      required:
      - product_model
      - brand
      - category
      - aspects
      - exemplar_mentions
      title: ProductAspectResponse
    ProductSummary:
      properties:
        brand:
          $ref: '#/components/schemas/Brand'
        category:
          $ref: '#/components/schemas/Category'
        product_model:
          type: string
          title: Product Model
        overall_score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Overall Score
        mention_count:
          type: integer
          title: Mention Count
      type: object
      required:
      - brand
      - category
      - product_model
      - overall_score
      - mention_count
      title: ProductSummary
    SentimentLabel:
      type: string
      enum:
      - positive
      - negative
      - neutral
      - mixed
      title: SentimentLabel
    ShareOfAspect:
      properties:
        brand:
          $ref: '#/components/schemas/Brand'
        aspect:
          type: string
          title: Aspect
        mention_share:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Mention Share
          description: Fraction of all mentions for this aspect
        avg_score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Avg Score
      type: object
      required:
      - brand
      - aspect
      - mention_share
      - avg_score
      title: ShareOfAspect
...
    SourcePlatform:
      type: string
      enum:
      - reddit
      - amazon
      - youtube
      - trustpilot
      - twitter
      - other
      title: SourcePlatform
    TimeseriesPoint:
      properties:
        date:
          type: string
          title: Date
          description: ISO date YYYY-MM-DD
        score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Score
        mention_count:
          type: integer
          title: Mention Count
      type: object
      required:
      - date
      - score
      - mention_count
      title: TimeseriesPoint
    TimeseriesResponse:
      properties:
        series:
          items:
            $ref: '#/components/schemas/TimeseriesPoint'
          type: array
          title: Series
        brand:
          anyOf:
          - $ref: '#/components/schemas/Brand'
          - type: 'null'
        category:
          anyOf:
          - $ref: '#/components/schemas/Category'
          - type: 'null'
      type: object
      required:
      - series
      title: TimeseriesResponse
    TopicCluster:
      properties:
        topic_id:
          type: string
          title: Topic Id
        label:
          type: string
          title: Label
        mention_count:
          type: integer
          title: Mention Count
        avg_score:
          type: number
          maximum: 1.0
          minimum: -1.0
          title: Avg Score
        momentum:
          type: number
          title: Momentum
          description: Recent growth rate
        exemplar_quotes:
          items:
            type: string
          type: array
          title: Exemplar Quotes
        trend:
          items:
            $ref: '#/components/schemas/TimeseriesPoint'
          type: array
          title: Trend
        is_novel:
          type: boolean
          title: Is Novel
          description: True when momentum > 0.3, mention_count < 25, and first_seen_at
            within last 14 days
          default: false
      type: object
      required:
      - topic_id
      - label
      - mention_count
      - avg_score
      - momentum
      - exemplar_quotes
      - trend
      title: TopicCluster
    TopicExplorerResponse:
      properties:
        topics:
          items:
            $ref: '#/components/schemas/TopicCluster'
          type: array
          title: Topics
      type: object
      required:
      - topics
      title: TopicExplorerResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError